@@ -1,22 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:kifu_viewer/configs/app_themes.dart';
-import 'package:kifu_viewer/localizations.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:kifu_viewer/theme/app_themes.dart';
 import 'package:kifu_viewer/widgets/home_screen/home_screen.dart';
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      localizationsDelegates: const [
-        AppLocalizationsDelegate(),
-        GlobalMaterialLocalizations.delegate,
-        GlobalWidgetsLocalizations.delegate,
-      ],
-      supportedLocales: AppLocalizationsDelegate.supportedLocals,
+      localizationsDelegates: AppLocalizations.localizationsDelegates,
+      supportedLocales: AppLocalizations.supportedLocales,
       theme: AppThemes.light,
       debugShowCheckedModeBanner: false,
       shortcuts: {
